% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_boxplot_by_group.R
\name{summary_and_boxplot_by_group}
\alias{summary_and_boxplot_by_group}
\title{Create summary statistics and boxplot across groups}
\usage{
summary_and_boxplot_by_group(
  data,
  group,
  var,
  drop_na = TRUE,
  alpha = 0.5,
  scale_y = FALSE
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{group}{The column containing different groups, which decides how the data will be grouped.}

\item{var}{A numeric vector. Summary statistics are computed based on it.}

\item{drop_na}{A logical value indicating whether NA values should be stripped before the computation proceeds.
If \code{TRUE}, NA values will be removed. If \code{FALSE}, NA values will be kept.
The default value is \code{TRUE}.}

\item{alpha}{A numeric value representing the transparency of the boxplot.
The default value is 0.5,}

\item{scale_y}{A logical value indicating whether the y-axis of the boxplot should be transformed to log10 scale.
If \code{TRUE}, y-axis will be scaled, otherwise \code{FALSE}. The default value is \code{FALSE}.}
}
\value{
A list containing (1) a tibble of summary statistics and, (2) the boxplot.
}
\description{
This function calculates the summary statistics (min, max, mean, median) of a variable in a data frame
across different groups. It also creates a boxplot where x-axis represents groups and y-axis represents the variable.
}
\examples{
summary_and_boxplot_by_group(datateachr::apt_buildings, property_type, no_of_units)
summary_and_boxplot_by_group(datateachr::apt_buildings, ward, no_of_storeys,
                             alpha = 0.3, scale_y = TRUE)

}
